refactor: remove obsolete migration files and initial schema
