+++ backend/alembic/versions/0f9103bfe8bf_initial_schema.py
@@ -1,0 +1,153 @@
+"""initial schema
+
+Revision ID: 0f9103bfe8bf
+Revises: 
+Create Date: 2025-09-19 19:40:27.664086
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '0f9103bfe8bf'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('modalities',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('category', sa.String(length=100), nullable=True),
+    sa.Column('currency', sa.String(length=3), nullable=False),
+    sa.Column('default_price_cents', sa.Integer(), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    op.create_table('professionals',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('email', sa.String(length=255), nullable=False),
+    sa.Column('full_name', sa.String(length=255), nullable=False),
+    sa.Column('phone', sa.String(length=20), nullable=True),
+    sa.Column('phone_country_code', sa.String(length=10), nullable=True),
+    sa.Column('phone_number', sa.String(length=20), nullable=True),
+    sa.Column('hashed_password', sa.String(length=255), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=True),
+    sa.Column('is_verified', sa.Boolean(), nullable=True),
+    sa.Column('profile_picture', sa.Text(), nullable=True),
+    sa.Column('license_number', sa.String(length=100), nullable=True),
+    sa.Column('years_experience', sa.Integer(), nullable=True),
+    sa.Column('rate_cents', sa.Integer(), nullable=False),
+    sa.Column('custom_rate_cents', sa.Integer(), nullable=True),
+    sa.Column('currency', sa.String(length=3), nullable=True),
+    sa.Column('bio', sa.Text(), nullable=True),
+    sa.Column('academic_experience', sa.Text(), nullable=True),
+    sa.Column('work_experience', sa.Text(), nullable=True),
+    sa.Column('certifications', sa.Text(), nullable=True),
+    sa.Column('languages', sa.ARRAY(sa.String()), nullable=True),
+    sa.Column('therapy_approaches_ids', sa.ARRAY(sa.String()), nullable=True),
+    sa.Column('specialty_ids', sa.ARRAY(sa.String()), nullable=True),
+    sa.Column('timezone', sa.String(length=50), nullable=True),
+    sa.Column('working_hours', sa.Text(), nullable=True),
+    sa.Column('emergency_contact', sa.String(length=255), nullable=True),
+    sa.Column('emergency_phone', sa.String(length=20), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_professionals_email'), 'professionals', ['email'], unique=True)
+    op.create_table('specialties',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('category', sa.String(length=100), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    op.create_table('therapeutic_approaches',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('category', sa.String(length=100), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name')
+    )
+    op.create_table('users',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('email', sa.String(length=255), nullable=False),
+    sa.Column('full_name', sa.String(length=255), nullable=False),
+    sa.Column('phone', sa.String(length=20), nullable=True),
+    sa.Column('hashed_password', sa.String(length=255), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=True),
+    sa.Column('is_verified', sa.Boolean(), nullable=True),
+    sa.Column('profile_picture', sa.Text(), nullable=True),
+    sa.Column('date_of_birth', sa.DateTime(), nullable=True),
+    sa.Column('emergency_contact', sa.String(length=255), nullable=True),
+    sa.Column('emergency_phone', sa.String(length=20), nullable=True),
+    sa.Column('preferences', sa.Text(), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+    op.create_table('professional_modalities',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('professional_id', sa.UUID(), nullable=False),
+    sa.Column('modality_id', sa.UUID(), nullable=False),
+    sa.Column('modality_name', sa.String(length=255), nullable=False),
+    sa.Column('virtual_price', sa.Integer(), nullable=False),
+    sa.Column('presencial_price', sa.Integer(), nullable=False),
+    sa.Column('offers_presencial', sa.Boolean(), nullable=False),
+    sa.Column('description', sa.String(length=1000), nullable=True),
+    sa.Column('is_default', sa.Boolean(), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.ForeignKeyConstraint(['modality_id'], ['modalities.id'], ),
+    sa.ForeignKeyConstraint(['professional_id'], ['professionals.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('professional_specialties',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('professional_id', sa.UUID(), nullable=False),
+    sa.Column('specialty_id', sa.UUID(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.ForeignKeyConstraint(['professional_id'], ['professionals.id'], ),
+    sa.ForeignKeyConstraint(['specialty_id'], ['specialties.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('professional_therapeutic_approaches',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('professional_id', sa.UUID(), nullable=False),
+    sa.Column('therapeutic_approach_id', sa.UUID(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('CURRENT_TIMESTAMP'), nullable=False),
+    sa.ForeignKeyConstraint(['professional_id'], ['professionals.id'], ),
+    sa.ForeignKeyConstraint(['therapeutic_approach_id'], ['therapeutic_approaches.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('professional_therapeutic_approaches')
+    op.drop_table('professional_specialties')
+    op.drop_table('professional_modalities')
+    op.drop_index(op.f('ix_users_email'), table_name='users')
+    op.drop_table('users')
+    op.drop_table('therapeutic_approaches')
+    op.drop_table('specialties')
+    op.drop_index(op.f('ix_professionals_email'), table_name='professionals')
+    op.drop_table('professionals')
+    op.drop_table('modalities')
+    # ### end Alembic commands ###
