--- backend/alembic/versions/backup/79edc6788fbd_add_specialties_table_and_professional_.py
@@ -1,42 +1,0 @@
-"""Add specialties table and professional specialty fields
-
-Revision ID: 79edc6788fbd
-Revises: 
-Create Date: 2025-09-12 17:05:16.228668
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '79edc6788fbd'
-down_revision = '8896a8b86d20'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('professionals', sa.Column('specialty_id', sa.UUID(), nullable=True))
-    op.add_column('professionals', sa.Column('custom_rate_cents', sa.Integer(), nullable=True))
-    op.alter_column('professionals', 'certifications',
-               existing_type=postgresql.ARRAY(sa.VARCHAR()),
-               type_=sa.Text(),
-               existing_nullable=True)
-    op.create_foreign_key(None, 'professionals', 'specialties', ['specialty_id'], ['id'])
-    op.drop_column('professionals', 'education')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('professionals', sa.Column('education', sa.TEXT(), autoincrement=False, nullable=True))
-    op.drop_constraint(None, 'professionals', type_='foreignkey')
-    op.alter_column('professionals', 'certifications',
-               existing_type=sa.Text(),
-               type_=postgresql.ARRAY(sa.VARCHAR()),
-               existing_nullable=True)
-    op.drop_column('professionals', 'custom_rate_cents')
-    op.drop_column('professionals', 'specialty_id')
-    # ### end Alembic commands ###
